refactor: Split mj-web-observer.ts into services and domain logic (#22)
